move upsample_yearly_data to resample
