+++ mesmer/__init__.py
@@ -20,6 +20,7 @@ from mesmer import (
     grid,
     io,
     mask,
+    resample,
     stats,
     testing,
     utils,
@@ -48,6 +49,7 @@ __all__ += [
     "get_options",
     "grid",
     "mask",
+    "resample",
     "set_options",
     "stats",
     "testing",
